# MM-54998: Optimize JSON marshalling in websocket broadcast