@agnivade 2023-11-08
Results look good when comparing flame graphs. But overall, there is no difference in the user count because the extra CPU headroom generated is just filled up by the cache serialization/deserialization and the CPU thresholds limit any further improvement.

Profile from a ceiling test: (I didn't get one from the master test, but it should be representative because it works in percentages)
![flame_graph](https://github.com/mattermost/mattermost/assets/1774000/e3a1731c-4690-4dd4-97eb-291513d3eef2)

Optimized:
![opt_ws_flame_profile](https://github.com/mattermost/mattermost/assets/1774000/0855d667-c817-4e9f-8846-3534213d542f)

You can see how the cache serialization starts to become more prominent in the new profile.